requirements.txt
pdfplumber
camelot-py[cv]
pytesseract
opencv-python
pandas
pdf2image
PyPDF2==2.12.1
